Multiple Inheritence
You are required to design two parent classes, InfoBasc and InfoDep. The description of the two classes are as follows 
a.	Class: InfoBasc: 
This class has 3 data members (Employer Name, ID, and gender), and one function member getInfo, which asks the employer about his name, ID number and gender.  
b.	Class: InfoDep 
This class has 3 data members (Department name, number of employees in the department, and the Department’s manager) and one function member getDepInfo, which asks the employer about this information  
After creating these 2 classes, you are required to create a class called employer which inherits from these 2 classes. It has 2-member function to get and display both the employer’s basic and department info. 
 
In the main function, create an object for the class employer and show how you can get and display the employer’s info.